+++ ThirdPartyAdapters/adaptertestkit/src/main/java/com/google/ads/mediation/adaptertestkit/AdErrorMatcher.kt
@@ -20,8 +20,6 @@ import org.mockito.ArgumentMatcher
 class AdErrorMatcher(val expected: AdError) : ArgumentMatcher<AdError> {
 
   override fun matches(actual: AdError): Boolean {
-    return actual.code == expected.code &&
-      actual.message == expected.message &&
-      actual.domain == expected.domain
+    return actual.code == expected.code && actual.domain == expected.domain
   }
 }